+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.3.30]
+
+- Released @ 12/2025 (UTC)
+- Minor tweaks
+
+## [0.3.29]
+
+- Released @ 12/2025 (UTC)
+- Add internalIndex to LogItem
+
+## [0.3.28]
+
+- Released @ 11/2025 (UTC)
+- feat: Add metadata field, add fixes
+
 ## [0.3.27]
 
 - Released @ 9/2025 (UTC)